requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
clamd==1.0.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.3
dj-database-url==2.3.0
Django==5.2
django-allauth==65.7.0
django-elasticsearch-dsl==8.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
elastic-transport==8.17.1
elasticsearch==8.18.1
elasticsearch-dsl==8.18.0
idna==3.10
kombu==5.5.3
lxml==5.4.0
onetimepass==1.0.1
phonenumbers==9.0.4
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
PyMuPDF==1.25.5
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
redis==6.0.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==12.1.0
totp==1.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
